Último commit WTForms

+++ appFlask/routes_main.py
@@ -35,8 +35,9 @@ def hello_world():
 
 @main_bp.route('/products/list')
 def item_list():
+    deleteForm = DeleteProductForm()
     items = db.session.query(Product)
-    return render_template('products/list.html', items=items)
+    return render_template('products/list.html', items=items, deleteForm=deleteForm)
 
 @main_bp.route('/products/create', methods=["GET", "POST"])
 def item_create():
@@ -68,43 +69,36 @@ def item_read(id):
 
 @main_bp.route('/products/edit/<int:id>', methods=["GET", "POST"])
 def item_edit(id):
+    cat = db.session.query(Category)
     info = db.session.query(Product).filter(Product.id == id).one()
-    if request.method == 'GET':
-        form = ProductForm()
-        cat = db.session.query(Category)
-        return render_template('products/create-edit.html', cat = cat, info = info, form=form)
-    elif request.method == 'POST':
-        data = request.form
-        id = '1'
-        if request.files["foto"]:
-            file = request.files["foto"]
-            if file and allowed_file(file.filename):
-                filename = secure_filename(file.filename)
-                file.save(os.path.join(UPLOAD_FOLDER, filename))
-                img = file.filename
-                remove(os.path.join(UPLOAD_FOLDER, request.form["photo"]))
-        else:
-                img = request.form["photo"]
-        info.title = data["nombre"]
-        info.description = data["descripcion"]
-        info.photo = img
-        info.price = float(data["precio"])
-        info.category_id = int(data["categoria"])
-        info.seller_id = int(id)
-        info.updated = now()
-        db.session.add(info)
+    form = ProductForm(obj = info)
+    form.category_id.choices=[(category.name) for category in cat]
+        
+    if request.method == 'GET' :
+        return render_template('products/create-edit.html', info=info, form = form)
+    elif request.method == 'POST' and form.validate_on_submit():
+        new_product = db.session.query(Product).filter(Product.id == id).one()
+        form.populate_obj(new_product)        
+        file = form.photo.data
+        if file and allowed_file(file.filename):
+            filename = secure_filename(file.filename)
+            new_product.photo = filename
+            file.save(os.path.join(UPLOAD_FOLDER, filename))
+        
+        new_product.updated = datetime.utcnow()
+
         db.session.commit()
-        return redirect(url_for('main_bp.item_list'))
+
+    return redirect(url_for('main_bp.item_list'))
 
 @main_bp.route('/products/delete/<int:id>', methods=["GET", "POST"])
 def item_delete(id):
     info = db.session.query(Product).filter(Product.id == id).one()
-    if request.method == 'GET':
-        form = DeleteProductForm()
-        return render_template('products/delete.html',info = info, form=form)
-    elif request.method == 'POST':
+    form = DeleteProductForm(obj = info)
+    if request.method == 'POST' and form.validate_on_submit():
         db.session.delete(info)
         db.session.commit()
-        data = request.form
-        remove(os.path.join(UPLOAD_FOLDER, data["foto"]))
+        remove(os.path.join(UPLOAD_FOLDER, info.photo))
         return redirect(url_for('main_bp.item_list'))
+    
+    return redirect(url_for('main_bp.item_list'))
